update A8 A9
