+++ S2SRL/SymbolicExecutor/symbolics_webqsp.py
@@ -14,39 +14,6 @@ app = Flask(__name__)
 # post_url = "http://10.201.34.3:5002/post"
 # # local server
 post_url = "http://127.0.0.1:5001/post"
-b_print = True
-
-
-class WebQSP(object):
-    def __init__(self, id, question, action_sequence_list, entity, relation, type, entity_mask, relation_mask, type_mask, mask_action_sequence_list, answerlist):
-        self.id = id
-        self.question = question
-        self.action_sequence_list = action_sequence_list
-        self.entity = entity
-        self.relation = relation
-        self.type = type
-        self.entity_mask = entity_mask
-        self.relation_mask = relation_mask
-        self.type_mask = type_mask
-        self.mask_action_sequence_list = mask_action_sequence_list
-        self.answerlist = answerlist
-
-    def obj_2_json(obj):
-        return {
-            obj.id: {
-                "question": obj.question,
-                "action_sequence_list": obj.action_sequence_list,
-                "entity": obj.entity,
-                "relation": obj.relation,
-                "type": obj.type,
-                "entity_mask": obj.entity_mask,
-                "relation_mask": obj.relation_mask,
-                "type_mask": obj.type_mask,
-                "mask_action_sequence_list": obj.mask_action_sequence_list,
-                "answers": obj.answerlist,
-            }
-        }
-
 
 class Symbolics_WebQSP():
 
@@ -693,535 +660,3 @@ class Action():
 
     def to_str(self):
         return "{{\'{0}\':[\'{1}\', \'{2}\', \'{3}\']}}".format(self.action_type, self.e, self.r, self.t)
-
-# parse sparql in dataset to action sequence
-def processSparql(sparql_str, id="empty"):
-        sparql_list = []
-        untreated_list = sparql_str.split("\n")
-        answer_keys = []
-        for untreated_str in untreated_list:
-            action_type = "A1"
-            s = ""
-            r = ""
-            t = ""
-
-            # remove note
-            note_index = untreated_str.find("#")
-            if note_index != -1:
-                untreated_str = untreated_str[0:note_index]
-            # remove /t
-            untreated_str = untreated_str.replace("\t", "")
-            untreated_str = untreated_str.strip()
-            if untreated_str == '':
-                continue
-
-            if "UNION" == untreated_str:
-                if b_print:
-                    print("has union", id)
-
-            if untreated_str.startswith("SELECT"):  # find answer key
-                for item in untreated_str.split(" "):
-                    if "?" in item:
-                        answer_keys.append(item.replace(" ", ""))
-            elif untreated_str.startswith("PREFIX") or "langMatches" in untreated_str:
-                # ignore
-                pass
-            elif untreated_str.startswith("FILTER (?x != ns:"):
-                # filter not equal
-                action_type = "A5"
-                s = "?x"
-                # s = "ANSWER"
-                t = untreated_str.replace("FILTER (?x != ns:", "").replace(")", "").replace(" ", "")
-                action_item = Action(action_type, s, r, t)
-                if isValidAction(action_item):
-                    sparql_list.append(action_item)
-            elif untreated_str.count("?") == 2 and ("FILTER" not in untreated_str or "EXISTS" not in untreated_str):
-                action_type = "A4"  # joint
-                triple_list = untreated_str.split(" ")
-                if len(triple_list) == 4:
-                    s = triple_list[0].replace("ns:", "")
-                    r = triple_list[1].replace("ns:", "")
-                    t = triple_list[2].replace("ns:", "")
-                    if s != "" and r != "" and t != "":
-                        action_item = Action(action_type, s, r, t)
-                        if isValidAction(action_item):
-                            sparql_list.append(action_item)
-            elif untreated_str.count("?") == 1 and untreated_str.startswith("ns:"):
-                # base action: select
-                action_type = "A1"
-                triple_list = untreated_str.split(" ")
-                if len(triple_list) == 4:
-                    s = triple_list[0].replace("ns:", "")
-                    r = triple_list[1].replace("ns:", "")
-                    t = triple_list[2].replace("ns:", "")
-                    if s != "" and r != "" and t != "":
-                        action_item = Action(action_type, s, r, t)
-                        if isValidAction(action_item):
-                            sparql_list.append(action_item)
-            elif untreated_str.count("?") == 1 and untreated_str.startswith("?"):
-                # ?x ns:a ns:b
-                # if have e,  A3 : filter variable: find sub set fits the bill
-                # if don't have e, A1_3 :find e
-                action_type = "A3"
-                triple_list = untreated_str.split(" ")
-                if True:
-                    s = triple_list[0].replace("ns:", "")
-                    r = triple_list[1].replace("ns:", "")
-                    t = triple_list[2].replace("ns:", "")
-                    if s != "" and r != "" and t != "":
-                        # for action in sparql_list:
-                        #     if action.e == s or action.t == s:  # already has variable
-                        #         action_type = "A4"
-                        # special for webqsp: swap s and t ,A3->A1 "6" means single action_seq
-                        # print(len(untreated_list), "length of untreated_list")
-                        # if len(untreated_list) == 6:
-                        #     action_type = "A1"
-                        #     temp = s
-                        #     s = t
-                        #     t = temp
-                        action_item = Action(action_type, s, r, t)
-                        if isValidAction(action_item):
-                            sparql_list.append(action_item)
-                # print(action_item)
-
-
-        # reorder list
-        reorder_sparql_list = reorder(sparql_list, answer_keys)
-        # reorder_sparql_list = sparql_list
-        # for astr in reorder_sparql_list:
-        #     print(astr.to_str())
-
-        old_sqarql_list = []
-        for item in reorder_sparql_list:
-            seqset = {}
-            seqlist = []
-            seqlist.append(item.e)
-            seqlist.append(item.r)
-            seqlist.append(item.t)
-            seqset[item.action_type] = seqlist
-            old_sqarql_list.append(seqset)
-        return old_sqarql_list
-
-# parse sparql for value type answer
-def processSparql_value(sparql_str, id="empty"):
-        sparql_list = []
-        untreated_list = sparql_str.split("\n")
-        answer_keys = []
-        value_len = len(untreated_list)
-        if value_len == 8:
-            value_str = untreated_list[5]
-            value_str_list = value_str.split(" ")
-            if len(value_str_list) == 4 and value_str_list[2] == '?x' and value_str_list[3] == '.':
-                s = value_str_list[0]
-                r = value_str_list[1]
-                return [s, r]
-        return []
-
-def isValidAction(action_item):
-    return (action_item.e.startswith("m.") or action_item.e.startswith("?"))\
-           and (action_item.t.startswith("m.") or action_item.t.startswith("?"))
-
-def reorder(sparql_list, answer_keys):
-    count = 0
-    final_len = len(sparql_list)
-    reorder_sparql_list = []
-
-    last_variable = ""
-    # while len(sparql_list) != 0:
-    for key in answer_keys:
-        # has_last_select = False
-        add_next_variable(sparql_list, key, reorder_sparql_list)
-        # print(reorder_sparql_list)
-            # contains key
-            # 提取包含key并排序
-            # for action_item in sparql_list:
-            #     # last select action of answer key
-            #     if action_item.t == key:
-            #         reorder_sparql_list.append(action_item)
-            #         sparql_list.remove(action_item)
-            #         break
-            #
-            #         if action_item.e.startswith('?'):
-            #             last_variable = action_item.e
-            #         has_last_select = True
-            #     if has_last_select:
-            #         seq = action_item.to_str()
-            #         print (seq)
-            #         if len(sparql_list) == 1:
-            #             reorder_sparql_list.append(action_item)
-            #             sparql_list.remove(action_item)
-            #             break
-            #         if "?x" in seq:
-            #             if seq.count("?") == 1:
-            #                 reorder_sparql_list.append(action_item)
-            #             elif seq.count("?") == 2:
-            #                 reorder_sparql_list.append(action_item)
-            #                 # define next variable
-            #                 next_variable = action_item.e if (action_item.t == "?x") else action_item.t
-            #             sparql_list.remove(action_item)
-            #             break
-            #         elif next_variable != "" and next_variable in seq:
-            #             if seq.count("?") == 1:
-            #                 reorder_sparql_list.append(action_item)
-            #             elif seq.count("?") == 2:
-            #                 reorder_sparql_list.append(action_item)
-            #                 # define next variable
-            #                 next_variable = action_item.e if (action_item.t == next_variable) else action_item.t
-            #             sparql_list.remove(action_item)
-            #             break
-    reorder_sparql_list.reverse()
-    return reorder_sparql_list
-
-def add_next_variable(sparql_list, variable_key, reorder_sparql_list):
-    for action_item in sparql_list:
-        # filter_not_equal is always the last action
-        if action_item.action_type == "A5":
-            reorder_sparql_list.append(action_item)
-            sparql_list.remove(action_item)
-            break
-
-    if variable_key == "":
-        return
-
-    variable_sql_list = []
-    for sql in sparql_list:
-        if variable_key == sql.e or variable_key == sql.t:
-            variable_sql_list.append(sql)
-
-    if len(variable_sql_list) == 0:
-        return
-    if len(variable_sql_list) == 1:
-        cur_action = variable_sql_list[0]
-        if cur_action == "?x" and "?" not in cur_action.t and cur_action.action_type == "A3":
-            cur_action.action_type = "A1"
-            e = cur_action.e
-            t = cur_action.t
-            cur_action.e = t
-            cur_action.t = e
-            reorder_sparql_list.append(sql)
-            sparql_list.remove(sql)
-            variable_sql_list.remove(sql)
-            return
-
-    next_variable = ""
-    for sql in variable_sql_list:
-        if sql.action_type == "A3":
-            reorder_sparql_list.append(sql)
-            sparql_list.remove(sql)
-            variable_sql_list.remove(sql)
-
-    for sql in variable_sql_list:
-        if sql.action_type == "A4":
-            # next_variable
-            next_variable = sql.e
-            reorder_sparql_list.append(sql)
-            sparql_list.remove(sql)
-            variable_sql_list.remove(sql)
-
-    for sql in variable_sql_list:
-        if sql.action_type == "A2":
-            reorder_sparql_list.append(sql)
-            sparql_list.remove(sql)
-            variable_sql_list.remove(sql)
-
-    for sql in variable_sql_list:
-        if sql.action_type == "A1":
-            reorder_sparql_list.append(sql)
-            sparql_list.remove(sql)
-            variable_sql_list.remove(sql)
-
-    return add_next_variable(sparql_list, next_variable, reorder_sparql_list)
-
-w_1 = 0.2
-def calc_01_reward(answer, true_answer):
-    true_reward = 0.0
-    try:
-        if len(true_answer) == 0:
-            if len(answer) == 0:
-                return 1.0
-            else:
-                return w_1
-        else:
-            right_count = 0
-            for e in answer:
-                if e in true_answer:
-                    right_count += 1
-            return float(right_count) / float(len(true_answer))
-    except:
-        return true_reward
-
-w_1 = 0.2
-def calc_01_reward_type(res_answer, true_answer, type = "f1"):
-    true_reward = 0.0
-    res_answer = set(res_answer)
-    true_answer = set(true_answer)
-    intersec = set(res_answer).intersection(set(true_answer))
-    if len(true_answer) == 0:
-        return 0.0
-    if type == "jaccard":
-        union = set([])
-        union.update(res_answer)
-        union.update(true_answer)
-        true_reward = float(len(intersec)) / float(len(union))
-    elif type == "recall":
-        true_reward = float(len(intersec)) / float(len(true_answer))
-    elif type == "f1":
-        if len(res_answer) == 0:
-            prec = 0.0
-        else:
-            prec = float(len(intersec)) / float(len(res_answer))
-        rec = float(len(intersec)) / float(len(true_answer))
-        if prec == 0 and rec == 0:
-            true_reward = 0.0
-        else:
-            true_reward = (2.0 * prec * rec) / (prec + rec)
-    return true_reward
-
-if __name__ == "__main__":
-
-    seq74 =[
-            {'A1': ['m.01cg88', 'type.object.name', '?x']},
-        ]
-    symbolic_exe = Symbolics_WebQSP(seq74)
-    answer = symbolic_exe.executor()
-    print("answer 74: ", answer)
-    #
-    # seq_3496 =[
-    #         # {'A1': ['m.0b90_r', 'location.statistical_region.places_exported_to', '?y']},
-    #         # {'A4': ['?y', 'location.imports_and_exports.exported_to', '?x']},
-    #         {'A1': ['m.0b90_r', 'location.statistical_region.places_imported_from', '?y']},
-    #         {'A4': ['?y', 'location.imports_and_exports.imported_from', '?x']},
-    #     ]
-    # symbolic_exe = Symbolics_WebQSP(seq_3496)
-    # answer = symbolic_exe.executor()
-    # print("answer WebQTrn_3496: ", answer)
-
-    # Load WebQuestions Semantic Parses
-    WebQSPList = []
-    WebQSPList_Correct = []
-    to_handle_list = []
-    WebQSPList_Incorrect = []
-    no_gold_answer = []
-    AnswerType_Value_idlist = []
-    to_add_list = []
-    result_list = []
-    no_x_list = []
-    json_errorlist = []
-    true_count = 0
-
-    # correct
-    with open("WebQSPList_Correct.json", "r", encoding='UTF-8') as correct_list:
-        WebQSPList_Correct = json.load(correct_list)
-    errorlist = []
-    with open("WebQSP.train.json", "r", encoding='UTF-8') as webQaTrain:
-        load_dictTrain = json.load(webQaTrain)
-    with open("WebQSP.test.json", "r", encoding='UTF-8') as webQaTest:
-        load_dictTest = json.load(webQaTest)
-    with open("to_handle.json", "r", encoding='UTF-8') as to_handle_file:
-        to_handle_list = json.load(to_handle_file)
-
-    train_questions = load_dictTrain["Questions"]
-    test_questions = load_dictTest["Questions"]
-    all_questions = train_questions + test_questions
-    small_questions = train_questions[0:9] + test_questions[0:9]
-
-    process_questions = all_questions
-    # total rewards
-    total_reward = 0
-    test_count = 0
-    total_reward_jaccard = 0
-    total_reward_precision = 0
-    total_reward_recall = 0
-
-    all_count = 0
-    for parse_q in process_questions:
-        question = parse_q["ProcessedQuestion"]
-        for q in parse_q["Parses"]:
-            id = q["ParseId"]
-            if id in WebQSPList_Correct or id in to_handle_list:
-                continue
-            all_count += 1
-            sparql = q["Sparql"]
-            reward = 0.0
-            answerList = q["Answers"]
-            if len(answerList) == 0:
-                print(id, "no gold answer")
-                no_gold_answer.append(id)
-            else:
-                Answers = []
-                for an in answerList:
-                    Answers.append(an['AnswerArgument'])
-                answer_type = answerList[0]['AnswerType']
-                if answer_type == "Value":
-                    AnswerType_Value_idlist.append(id)
-                    sr_list = processSparql_value(sparql)
-                    if len(sr_list) == 2:
-                        for an in Answers:
-                            to_add_list.append(str(sr_list[0] + '**' + sr_list[1] + '**' + an))
-                else:
-                    # assert answer_type == "Entity"
-                    # continue
-                    # if id == "WebQTrn-193.P0" or id == "WebQTrn-194.P0":  # test one
-                    if True:    # test all
-                        # test seq
-                        true_answer = Answers
-                        test_sparql = sparql
-                        seq = processSparql(test_sparql, id)
-                        if b_print:
-                            print(id)
-                            print(len(seq))
-                        symbolic_exe = Symbolics_WebQSP(seq)
-                        answer = symbolic_exe.executor()
-                        if b_print:
-                            print("answer: ", answer)
-                            print("true_answer: ", true_answer)
-                        try:
-                            key = "?x"
-                            if key in answer:
-                                res_answer = answer[key]
-                                reward = calc_01_reward_type(res_answer, true_answer, "f1")
-                                if b_print:
-                                    print(id, reward)
-                                result_list.append({id: [seq, reward]})
-                                if reward != 1.0:
-                                    result_list.append({id: list(res_answer)})
-                                    result_list.append({id: list(true_answer)})
-
-                                reward_jaccard = calc_01_reward_type(res_answer, true_answer, "jaccard")
-                                reward_recall = calc_01_reward_type(res_answer, true_answer, "recall")
-                                reward_precision = calc_01_reward_type(res_answer, true_answer, "precision")
-                                test_count += 1
-                                if reward == 1.0:
-                                    # if get right answer, generate action sequence
-                                    true_count += 1
-                                    entity = set()
-                                    relation = set()
-                                    type = set()
-                                    e_index = 1
-                                    r_index = 1
-                                    t_index = 1
-                                    for srt in seq:
-                                        for k, v in srt.items():
-                                            if v[0] != "":
-                                                entity.add(v[0])
-                                            if v[1] != "":
-                                                relation.add(v[1])
-                                            if v[2] != "":
-                                                type.add(v[2])
-                                    entity = list(entity)
-                                    relation = list(relation)
-                                    type = list(type)
-                                    entity_mask = dict()
-                                    relation_mask = dict()
-                                    type_mask = dict()
-                                    for e in entity:
-                                        dict_entity = {e: "ENTITY{0}".format(e_index)}
-                                        entity_mask.update(dict_entity)
-                                        e_index += 1
-                                    for r in relation:
-                                        dict_relation = {r: "RELATION{0}".format(r_index)}
-                                        relation_mask.update(dict_relation)
-                                        r_index += 1
-                                    for t in type:
-                                        dict_type = {t: "TYPE{0}".format(t_index)}
-                                        type_mask.update(dict_type)
-                                        t_index += 1
-                                    mask_action_sequence_list = []
-
-                                    for srt in seq:
-                                        mask_set = {}
-                                        masklist = []
-                                        a_mask = ""
-                                        e_mask = ""
-                                        r_mask = ""
-                                        t_mask = ""
-                                        for k,v in srt.items():
-                                            a_mask = k
-                                            e_mask_key = v[0]
-                                            r_mask_key = v[1]
-                                            t_mask_key = v[2]
-                                            e_mask = entity_mask[e_mask_key] if e_mask_key != "" else ""
-                                            r_mask = relation_mask[r_mask_key] if r_mask_key != "" else ""
-                                            t_mask = type_mask[t_mask_key] if t_mask_key != "" else ""
-                                        if a_mask != "":
-                                            masklist.append(e_mask)
-                                            masklist.append(r_mask)
-                                            masklist.append(t_mask)
-                                            mask_set = {a_mask : masklist}
-                                            mask_action_sequence_list.append(mask_set)
-                                    if id != "" and question != "" and seq != "":
-                                        correct_item = WebQSP(id, question, seq, entity, relation, type,entity_mask,
-                                                              relation_mask, type_mask, mask_action_sequence_list, answerList)
-                                    # print(question)
-                                    # print(answer)
-                                    WebQSPList_Correct.append(id)
-                                else:
-                                    if b_print:
-                                        print('incorrect!', reward)
-                                        print("answer", answer)
-                                        print("seq", seq)
-                                        print("true_answer", true_answer)
-                                        print("id", id)
-                                        print(" ")
-                                    WebQSPList_Incorrect.append(id)
-                                    errorlist.append(id)
-                                    json_errorlist.append(q)
-
-                                total_reward += reward
-                                total_reward_jaccard += reward_jaccard
-                                total_reward_recall += reward_recall
-                                total_reward_precision += reward_precision
-                            else:
-                                no_x_list.append(id)
-                        except Exception as exception:
-                            print(exception)
-                            pass
-
-    print('all_count', all_count)
-    questions_count = len(process_questions)
-    mean_reward_jaccard = total_reward_jaccard / questions_count
-    mean_reward_recall = total_reward_recall / questions_count
-    mean_reward_precision = total_reward_precision / questions_count
-    mean_reward = total_reward / questions_count
-    print("mean_reward_jaccard: ", mean_reward_jaccard)
-    print("mean_reward_recall: ", mean_reward_recall)
-    print("mean_reward_precision: ", mean_reward_precision)
-    print("mean_reward_f1: ", mean_reward)
-    print("{0} pairs correct".format(true_count))
-    print(errorlist)
-
-    # not x
-    jsondata = json.dumps(no_x_list, indent=1)
-    fileObject = open('no_x_list.json', 'w')
-    fileObject.write(jsondata)
-    fileObject.close()
-
-    # result_list
-    jsondata = json.dumps(result_list, indent=1)
-    fileObject = open('result_list.json', 'w')
-    fileObject.write(jsondata)
-    fileObject.close()
-
-    # no_gold_answer
-    jsondata = json.dumps(no_gold_answer, indent=1)
-    fileObject = open('no_gold_answer.json', 'w')
-    fileObject.write(jsondata)
-    fileObject.close()
-
-    # AnswerType_Value_idlist
-    jsondata = json.dumps(AnswerType_Value_idlist, indent=1)
-    fileObject = open('AnswerType_Value_idlist.json', 'w')
-    fileObject.write(jsondata)
-    fileObject.close()
-
-    # to_add_list
-    jsondata = json.dumps(to_add_list, indent=1)
-    fileObject = open('to_add_list.json', 'w')
-    fileObject.write(jsondata)
-    fileObject.close()
-
-    # jsondata = json.dumps(WebQSPList_Correct, indent=1, default=WebQSP.obj_2_json)
-    jsondata = json.dumps(WebQSPList_Correct, indent=1)
-    fileObject = open('WebQSPList_Correct1.json', 'w')
-    fileObject.write(jsondata)
-    fileObject.close()
