--- data/webquestionssp/read.me
@@ -1,1 +1,0 @@
-The dataset is used to store triples in freebase which are used for webquestionssp.
